Update en.json
I looked through the English locale and found several small mistakes (misspellings), so I fixed them. I also changed "Did not" to "Didn't" one time, even thought it is not a mistake, because there is another "Didn't" in the text - just to make sure everything conforms to the same style. If this was useful for you, please add the hacktoberfest-accepted label. If not, no worries.
Thank you and don't hesitate to reach out in case you have any questions.